The new commits include the addition of two new files: "Project_Notebook.ipynb" and "data/StudentsPerformance.csv," as well as the renaming of the file "help.txt" to "helper_git/help.txt." These changes were made to the "main" branch of the repository. No changes were made to any other branches.
The "Project_Notebook.ipynb" file contains a Jupyter Notebook that will be used for a data analysis project. The "data/StudentsPerformance.csv" file contains data that will be used in the analysis. Finally, the "helper_git/help.txt" file was moved to a subdirectory "helper_git" to keep the main directory clean.

This commit message serves as a record of the changes made to the repository and helps to keep track of the history of the project.